Apply proposed changes
